Rename cliquet.logging module to cliquet.logs
